requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
Click==7.0
coverage==5.0.2
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-service-objects==0.5.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
isort==4.3.21
mccabe==0.6.1
pathspec==0.7.0
pkg-resources==0.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
regex==2019.12.20
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
whitenoise==5.0.1
